requirements.txt
scapy
paramiko
pyzmq
python-nmap
ldap3
requests